refactor: unify process instance incident queries into one hook
