+++ operate/client/src/modules/queries/incidents/useGetIncidentsByProcessInstance.ts
@@ -6,44 +6,46 @@
  * except in compliance with the Camunda License 1.0.
  */
 
+import type {QueryIncidentsResponseBody} from '@camunda/camunda-api-zod-schemas/8.8';
 import {useQuery} from '@tanstack/react-query';
-
 import {searchIncidentsByProcessInstance} from 'modules/api/v2/incidents/searchIncidentsByProcessInstance';
-const INCIDENTS_SEARCH_QUERY_KEY = 'incidentsSearch';
 
-const useGetIncidentsByProcessInstance = (
+const INCIDENTS_SEARCH_QUERY_KEY = 'incidentsSearchByProcessInstance';
+
+type QueryOptions<T> = {
+  enabled?: boolean;
+  enablePeriodicRefetch?: boolean;
+  select?: (result: QueryIncidentsResponseBody) => T;
+};
+
+const useGetIncidentsByProcessInstance = <T = QueryIncidentsResponseBody>(
   processInstanceKey: string,
-  elementInstanceKey?: string,
-  shouldFilterByElementInstance?: boolean,
-  options: {enabled: boolean} = {enabled: true},
+  options?: QueryOptions<T>,
 ) => {
   return useQuery({
-    queryKey: [
-      INCIDENTS_SEARCH_QUERY_KEY,
-      processInstanceKey,
-      elementInstanceKey,
-    ],
+    queryKey: [INCIDENTS_SEARCH_QUERY_KEY, processInstanceKey],
+    refetchInterval: () => (options?.enablePeriodicRefetch ? 5000 : false),
+    enabled: options?.enabled ?? !!processInstanceKey,
+    select: options?.select,
     queryFn: async () => {
       const {response, error} = await searchIncidentsByProcessInstance({
         processInstanceKey,
       });
-
-      if (response !== null) {
-        const incidents =
-          shouldFilterByElementInstance && elementInstanceKey
-            ? response.items.filter(
-                (incident) =>
-                  incident.elementInstanceKey === elementInstanceKey,
-              )
-            : response.items;
-
-        return incidents.length > 0 ? incidents : null;
+      if (error) {
+        throw error;
       }
 
-      throw error;
+      return response;
     },
-    ...options,
   });
 };
 
-export {useGetIncidentsByProcessInstance};
+function useProcessInstanceIncidentsCount(processInstanceKey: string): number {
+  const {data} = useGetIncidentsByProcessInstance(processInstanceKey, {
+    select: (incidents) => incidents.page.totalItems,
+  });
+
+  return data ?? 0;
+}
+
+export {useGetIncidentsByProcessInstance, useProcessInstanceIncidentsCount};
